pj1/requirements.txt
Flask
fastapi
flask-login
flask-wtf 
flask-sqlalchemy
flask-bcrypt
uvicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-multipart
gunicorn
Flask-Login
WTForms
python-dotenv

